Do not generate application layout file
As mentioned in https://github.com/thoughtbot/clearance/issues/701,
(which this solves), adding the application layout file to the list of
views managed by Clearance can lead to unexpected/adverse effects.

While it is true that including the file with the suggested
implementation of flashes and sign in/sign out logic is valuable in
providing both a clear example of what is necessary to implement the
library and in a good out-of-the-box experience, it can lead to
inadvertently overriding the existing layout or, like in the issue
mentioned above, have a surprising destructive effect when relying on
Rails generators to undo the views being copied.

Also, the very same implementation suggestion is presented to Stdout
when running the installer, so this change does not imply the loss of
that helpful input.

The PR moves the layout file to the testing application
`spec/app_templates/testapp/app/views/layouts/application.html.erb` in
order to maintain the acceptance tests passing as expected, and removes
`application.html.erb` from the list of files managed under the
Clearance generator.